requirements.txt
pandas
chromadb
sentence-transformers
uuid
groq
python-dotenv
Flask